feat: added update endpoint

+++ src/main/java/com/stockManagement/stockManagement/dto/request/ProductChangeRequestDto.java
@@ -2,5 +2,5 @@ package com.stockManagement.stockManagement.dto.request;
 
 import java.math.BigDecimal;
 
-public record ProductChangeRequestDto(Integer id, String name, Integer quantity, BigDecimal price) {
+public record ProductChangeRequestDto(Long id, String name, Integer quantity, BigDecimal price) {
 }
